+++ lib/fluidaudio_dart.dart
@@ -4,6 +4,7 @@
 library;
 
 export 'src/asr.dart' show FluidAsr;
+export 'src/ctc_vocabulary.dart' show FluidCtcVocabulary, FluidVocabularyTerm;
 export 'src/diarizer.dart'
     show
         FluidChunkEmbedding,
@@ -16,6 +17,7 @@ export 'src/eou.dart' show EouChunkSize, FluidEou;
 export 'src/events.dart' show FluidDownloadProgressFailure;
 export 'src/exceptions.dart'
     show FluidAudioException, FluidDownloadException, FluidInstanceGoneException;
+export 'src/itn.dart' show FluidItn;
 export 'src/models.dart' show FluidModels;
 export 'src/streaming_asr.dart' show FluidStreamingAsr;
 export 'src/system.dart' show FluidAudioSystem, FluidDebugEvent, FluidSystemInfo;